feat: Add a `ty::View`
`ty::View<'tcx, T>` acts like a `T` but internally stores a `Ty<'tcx>`
pointer. Thanks to this, it is possible to retrieve the original `Ty`
value without needing to ask the interner for it.

Looking up the already created type (`T`) takes a good chunk of the
time in `infer/outlives/verify.rs` so this should be a good speedup.
It may be applicable in other places as well, but those are far lower
when profiling.